Struct ethjson::spec::authority_round::AuthorityRoundParams

pub struct AuthorityRoundParams {
    pub step_duration: Uint,
    pub validators: ValidatorSet,
    pub start_step: Option<Uint>,
    pub validate_score_transition: Option<Uint>,
    pub validate_step_transition: Option<Uint>,
    pub immediate_transitions: Option<bool>,
    pub block_reward: Option<Uint>,
    pub block_reward_contract_transition: Option<Uint>,
    pub block_reward_contract_address: Option<Address>,
    pub maximum_uncle_count_transition: Option<Uint>,
    pub maximum_uncle_count: Option<Uint>,
    pub empty_steps_transition: Option<Uint>,
    pub maximum_empty_steps: Option<Uint>,
}

Authority params deserialization.

Fields
step_duration: Uint

Block duration, in seconds.

validators: ValidatorSet

Valid authorities

start_step: Option<Uint>

Starting step. Determined automatically if not specified. To be used for testing only.

validate_score_transition: Option<Uint>

Block at which score validation should start.

validate_step_transition: Option<Uint>

Block from which monotonic steps start.

immediate_transitions: Option<bool>

Whether transitions should be immediate.

block_reward: Option<Uint>

Reward per block in wei.

block_reward_contract_transition: Option<Uint>

Block at which the block reward contract should start being used.

block_reward_contract_address: Option<Address>

Block reward contract address (setting the block reward contract overrides the static block reward definition).

maximum_uncle_count_transition: Option<Uint>

Block at which maximum uncle count should be considered.

maximum_uncle_count: Option<Uint>

Maximum number of accepted uncles.

empty_steps_transition: Option<Uint>

Block at which empty step messages should start.

maximum_empty_steps: Option<Uint>

Maximum number of accepted empty steps.