Struct ethjson::spec::params::Params

pub struct Params {

    pub account_start_nonce: Option<Uint>,
    pub maximum_extra_data_size: Uint,
    pub min_gas_limit: Uint,
    pub network_id: Uint,
    pub chain_id: Option<Uint>,
    pub subprotocol_name: Option<String>,
    pub fork_block: Option<Uint>,
    pub fork_hash: Option<H256>,
    pub eip150_transition: Option<Uint>,
    pub eip160_transition: Option<Uint>,
    pub eip161abc_transition: Option<Uint>,
    pub eip161d_transition: Option<Uint>,
    pub eip98_transition: Option<Uint>,
    pub eip155_transition: Option<Uint>,
    pub validate_chain_id_transition: Option<Uint>,
    pub validate_receipts_transition: Option<Uint>,
    pub eip140_transition: Option<Uint>,
    pub eip210_transition: Option<Uint>,
    pub eip210_contract_address: Option<Address>,
    pub eip210_contract_code: Option<Bytes>,
    pub eip210_contract_gas: Option<Uint>,
    pub eip211_transition: Option<Uint>,
    pub eip145_transition: Option<Uint>,
    pub eip214_transition: Option<Uint>,
    pub eip658_transition: Option<Uint>,
    pub eip1052_transition: Option<Uint>,
    pub eip1283_transition: Option<Uint>,
    pub eip1283_disable_transition: Option<Uint>,
    pub eip1283_reenable_transition: Option<Uint>,
    pub eip1014_transition: Option<Uint>,
    pub eip1706_transition: Option<Uint>,
    pub eip1344_transition: Option<Uint>,
    pub eip1884_transition: Option<Uint>,
    pub eip2028_transition: Option<Uint>,
    pub eip2315_transition: Option<Uint>,
    pub eip2929_transition: Option<Uint>,
    pub eip2930_transition: Option<Uint>,
    pub eip1559_transition: Option<Uint>,
    pub eip3198_transition: Option<Uint>,
    pub eip3529_transition: Option<Uint>,
    pub eip3541_transition: Option<Uint>,
    pub eip3607_transition: Option<Uint>,
    pub dust_protection_transition: Option<Uint>,
    pub nonce_cap_increment: Option<Uint>,
    pub remove_dust_contracts: Option<bool>,
    pub gas_limit_bound_divisor: Uint,
    pub registrar: Option<Address>,
    pub apply_reward: Option<bool>,
    pub node_permission_contract: Option<Address>,
    pub max_code_size: Option<Uint>,
    pub max_transaction_size: Option<Uint>,
    pub max_code_size_transition: Option<Uint>,
    pub transaction_permission_contract: Option<Address>,
    pub transaction_permission_contract_transition: Option<Uint>,
    pub wasm_activation_transition: Option<Uint>,
    pub wasm_disable_transition: Option<Uint>,
    pub kip4_transition: Option<Uint>,
    pub kip6_transition: Option<Uint>,
    pub eip1559_base_fee_max_change_denominator: Option<Uint>,
    pub eip1559_elasticity_multiplier: Option<Uint>,
    pub eip1559_base_fee_initial_value: Option<Uint>,
    pub eip1559_base_fee_min_value: Option<Uint>,
    pub eip1559_base_fee_min_value_transition: Option<Uint>,
    pub eip1559_fee_collector: Option<Address>,
    pub eip1559_fee_collector_transition: Option<Uint>,
    pub validate_service_transactions_transition: Option<Uint>,

}

Spec params.

Fields
account_start_nonce: Option<Uint>

Account start nonce, defaults to 0.

maximum_extra_data_size: Uint

Maximum size of extra data.

min_gas_limit: Uint

Minimum gas limit.

network_id: Uint

Network id.

chain_id: Option<Uint>

Chain id.

subprotocol_name: Option<String>

Name of the main (“eth”) subprotocol.

fork_block: Option<Uint>

Option fork block number to check.

fork_hash: Option<H256>

Expected fork block hash.

eip150_transition: Option<Uint>

See main EthashParams docs.

eip160_transition: Option<Uint>

See main EthashParams docs.

eip161abc_transition: Option<Uint>

See main EthashParams docs.

eip161d_transition: Option<Uint>

See main EthashParams docs.

eip98_transition: Option<Uint>

See CommonParams docs.

eip155_transition: Option<Uint>

See CommonParams docs.

validate_chain_id_transition: Option<Uint>

See CommonParams docs.

validate_receipts_transition: Option<Uint>

See CommonParams docs.

eip140_transition: Option<Uint>

See CommonParams docs.

eip210_transition: Option<Uint>

See CommonParams docs.

eip210_contract_address: Option<Address>

See CommonParams docs.

eip210_contract_code: Option<Bytes>

See CommonParams docs.

eip210_contract_gas: Option<Uint>

See CommonParams docs.

eip211_transition: Option<Uint>

See CommonParams docs.

eip145_transition: Option<Uint>

See CommonParams docs.

eip214_transition: Option<Uint>

See CommonParams docs.

eip658_transition: Option<Uint>

See CommonParams docs.

eip1052_transition: Option<Uint>

See CommonParams docs.

eip1283_transition: Option<Uint>

See CommonParams docs.

eip1283_disable_transition: Option<Uint>

See CommonParams docs.

eip1283_reenable_transition: Option<Uint>

See CommonParams docs.

eip1014_transition: Option<Uint>

See CommonParams docs.

eip1706_transition: Option<Uint>

See CommonParams docs.

eip1344_transition: Option<Uint>

See CommonParams docs.

eip1884_transition: Option<Uint>

See CommonParams docs.

eip2028_transition: Option<Uint>

See CommonParams docs.

eip2315_transition: Option<Uint>

See CommonParams docs.

eip2929_transition: Option<Uint>

See CommonParams docs.

eip2930_transition: Option<Uint>

See CommonParams docs.

eip1559_transition: Option<Uint>

See CommonParams docs.

eip3198_transition: Option<Uint>

See CommonParams docs.

eip3529_transition: Option<Uint>

See CommonParams docs.

eip3541_transition: Option<Uint>

See CommonParams docs.

eip3607_transition: Option<Uint>

See CommonParams docs.

dust_protection_transition: Option<Uint>

See CommonParams docs.

nonce_cap_increment: Option<Uint>

See CommonParams docs.

remove_dust_contracts: Option<bool>

See CommonParams docs.

gas_limit_bound_divisor: Uint

See CommonParams docs.

registrar: Option<Address>

See CommonParams docs.

apply_reward: Option<bool>

Apply reward flag

node_permission_contract: Option<Address>

Node permission contract address.

max_code_size: Option<Uint>

See main EthashParams docs.

max_transaction_size: Option<Uint>

Maximum size of transaction RLP payload.

max_code_size_transition: Option<Uint>

See main EthashParams docs.

transaction_permission_contract: Option<Address>

Transaction permission contract address.

transaction_permission_contract_transition: Option<Uint>

Block at which the transaction permission contract should start being used.

wasm_activation_transition: Option<Uint>

Wasm activation block height, if not activated from start

wasm_disable_transition: Option<Uint>

Wasm deactivation block height, if activated.

kip4_transition: Option<Uint>

KIP4 activiation block height.

kip6_transition: Option<Uint>

KIP6 activiation block height.

eip1559_base_fee_max_change_denominator: Option<Uint>

Base fee max change denominator

eip1559_elasticity_multiplier: Option<Uint>

Elasticity multiplier

eip1559_base_fee_initial_value: Option<Uint>

Default value for the block base fee

eip1559_base_fee_min_value: Option<Uint>

Min value for the block base fee.

eip1559_base_fee_min_value_transition: Option<Uint>

Block at which the min value for the base fee starts to be used.

eip1559_fee_collector: Option<Address>

Address where EIP-1559 burnt fee will be accrued to.

eip1559_fee_collector_transition: Option<Uint>

Block at which the fee collector should start being used.

validate_service_transactions_transition: Option<Uint>

Block at which zero gas price transactions start being checked with Certifier contract.