Struct ethjson::vm::vm::Vm

pub struct Vm {
    pub calls: Option<Vec<Call>>,
    pub env: Env,
    pub transaction: Transaction,
    pub gas_left: Option<Uint>,
    pub logs: Option<H256>,
    pub output: Option<Bytes>,
    pub post_state: Option<State>,
    pub pre_state: State,
}

Represents vm execution environment before and after execution of transaction.

Fields
calls: Option<Vec<Call>>

Contract calls made internaly by executed transaction.

env: Env

Env info.

transaction: Transaction

Executed transaction

gas_left: Option<Uint>

Gas left after transaction execution.

logs: Option<H256>

Hash of logs created during execution of transaction.

output: Option<Bytes>

Transaction output.

post_state: Option<State>

Post execution vm state.

pre_state: State

Pre execution vm state.

Methods

impl Vm

pub fn out_of_gas(&self) -> bool

Returns true if transaction execution run out of gas.