Struct ethjson::vm::vm::Vm

pub struct Vm {
    pub calls: Option<Vec<Call>>,
    pub env: Env,
    pub transaction: Transaction,
    pub gas_left: Option<Uint>,
    pub logs: Option<Vec<Log>>,
    pub output: Option<Bytes>,
    pub post_state: Option<State>,
    pub pre_state: State,
}

Reporesents vm execution environment before and after exeuction of transaction.

Fields

calls

Contract calls made internaly by executed transaction.

env

Env info.

transaction

Executed transaction

gas_left

Gas left after transaction execution.

logs

Logs created during execution of transaction.

output

Transaction output.

post_state

Post execution vm state.

pre_state

Pre execution vm state.

Methods

impl Vm

fn out_of_gas(&self) -> bool

Returns true if transaction execution run out of gas.